+++ ayaka/context.py
@@ -1,8 +1,8 @@
 '''上下文'''
 import asyncio
-from typing import TYPE_CHECKING
-from sqlmodel import Session
-from .helpers import simple_repr, ContextGroup, Field
+from sqlmodel import Session as DBSession
+from typing import TYPE_CHECKING, Optional
+from .helpers import ContextGroup, Field
 from .adapters import AyakaEvent
 
 if TYPE_CHECKING:
@@ -24,29 +24,29 @@ class AyakaContext(ContextGroup):
     trigger: "AyakaTrigger"
     '''当前触发器'''
 
-    cmd: str
+    prefix: Optional[str] = None
+    '''当前命令前缀'''
+
+    cmd: str = ""
     '''当前命令'''
 
-    arg: str
+    arg: str = ""
     '''去掉命令的剩余文字'''
 
-    args: list[str]
+    args: list[str] = []
     '''剩余文字根据separate分割'''
 
-    nums: list[int]
+    nums: list[int] = []
     '''数字'''
 
     wait_tasks: list[asyncio.Task] = []
     '''数据库会话关闭前，会等待该队列中的所有任务结束'''
 
-    db_session: Session = Field(default_factory=get_db_session)
+    db_session: DBSession = Field(default_factory=get_db_session)
     '''数据库会话，请在trigger存在的时候使用'''
 
     db_session_flag: bool = False
     '''是否使用了db_session'''
-
-    def __repr__(self) -> str:
-        return simple_repr(self)
 
 
 ayaka_context = AyakaContext()